Brief description of the repo.
This repo contains a simple, automated version of the simple card game "WAR". From this code, we can, of course, add subroutine(s) that allow us to play out a tie, and we can give it a nice GUI so that people can have a good time using and interacting with it.